backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0
reportlab==4.2.5
PyPDF2==3.0.1
jinja2==3.1.5
aiofiles==24.1.0
python-dateutil==2.9.0.post0
